shared library support
When generating a TAB file for the application, elf2tab will search the
shared library object files for any shared library functions that the
application references. elf2tab generates a TOML file to describe the
functions that need references to be relocated in the application
binary. The TOML file is bundled into the TAR of the TAB for Tockloader
to use while it patches up the application binary.